Unindo as classes de tabelas em uma unica classe Singleton, e removendo o escopo das listas das tabelas
