+++ Cod3rsGrowth.Teste/TesteServicoFilme.cs
@@ -2,10 +2,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Cod3rsGrowth.Dominio;
 using Cod3rsGrowth.Servicos.Servicos;
 using Cod3rsGrowth.Servicos.Interfaces;
